requirements.txt
asgiref==3.5.2
attrs==21.4.0
dj-database-url==0.5.0
Django==4.0.6
django-environ==0.9.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
inflection==0.5.1
jsonschema==4.6.1
psycopg2-binary==2.9.3
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
